requirements.txt
langchain
langchain-ollama
langfuse
python-dotenv